requirement.txt
Flask
pydub
python-slugify
pytube
urllib3
gunicorn
flask_sqlalchemy
requests
sqlalchemy